third_party/WebKit/LayoutTests/typedcssom/computedstyle/line-height.html

<!DOCTYPE html>
<html>
<head>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
</head>
<body>
<div id='testElement'></div>

<script>

testElement.style.lineHeight = '10px';
testElement.style.fontSize = '100px';

var computedStyleMap = getComputedStyleMap(testElement);

test(function() {
  var result = computedStyleMap.get('line-height');
  assert_true(result instanceof CSSSimpleLength);
  assert_equals(result.cssText, '10px');
}, 'Getting a 10px lineHeight results in a CSSSimpleLength');

test(function() {
  testElement.style.lineHeight = '20px';
  var result = computedStyleMap.getAll('line-height');
  assert_equals(result.length, 1);
  assert_equals(result[0].cssText, '20px');
}, 'getAll for lineHeight returns a single value');

test(function() {
  testElement.style.lineHeight = '10%';
  var result = computedStyleMap.get('line-height');
  assert_true(result instanceof CSSSimpleLength);
  assert_equals(result.cssText, '10px');
}, 'Getting a 10% lineHeight results in a CSSSimpleLength');

test(function() {
  testElement.style.lineHeight = '0.2';
  var result = computedStyleMap.get('line-height');
  assert_true(result instanceof CSSSimpleLength);
  assert_equals(result.cssText, '20px');
}, 'Getting a number lineHeight results in a CSSSimpleLength');

test(function() {
  testElement.style.lineHeight = 'calc(10px + 10%)';
  var result = computedStyleMap.get('line-height');
  assert_true(result instanceof CSSSimpleLength);
  assert_equals(result.cssText, '20px');
}, 'Getting a calc lineHeight results in a CSSSimpleLength');

test(function() {
  testElement.style.lineHeight = 'normal';
  var result = computedStyleMap.get('line-height');
  assert_true(result instanceof CSSKeywordValue);
  assert_equals(result.cssText, 'normal');
}, 'Getting a normal lineHeight results in a CSSKeywordValue');

</script>
</body>
</html>
